Change create_slug to work in both Python 2.7 & 3.2.
This is pretty hackish and needs to be revisited.

Using markupsafe in place of xml.sax.saxutils handles a wider range of
HTML entities, and ensures a Unicode result in Python 2. Markupsafe is
already a dependency via Mako.

The six library will provide a better handling of the result of the
unicodedata operation. However the slug should probably remain Unicode
at this point and only be encoded to UTF-8 when it is being used at
the file system interface.